Move add-query-args to lib/url.
