+++ client/lib/route/index.ts
@@ -1,5 +1,5 @@
 export * from './path';
-export { default as addQueryArgs } from './add-query-args';
+export { default as addQueryArgs } from 'lib/url/add-query-args';
 export { default as normalize } from './normalize';
 export { default as trailingslashit } from './trailingslashit';
 export { default as untrailingslashit } from './untrailingslashit';
